Avoid implicit encoding upgrading in _resolve
When /frame?url=... is given a URL of a page which is not encoded with
latin-1 and contains a character reference whose code point is greater
than 255 in an attribute, a server will fail to return a response or
give a broken HTML.

Example:
    <html>
      <head><meta charset="utf-8"></head>
      <body><input value="&#x2713;">ほげ</body>
    </html>

HTML::Parser, which is used in HTML::ResolveLink, decodes character
references in attributes by default. If its code point is greater than
255, "UTF8 flag" is on. In concatenation of a UTF8-flagged string and
non-UTF8-flagged one, Perl implicitly decodes non-UTF8-flagged one
assuming it is encoded with latin-1. If it is not encoded with latin-1,
the result of this implicit decoding (so-called upgrading) will be a
broken string.

And the concatenated string is flagged as UTF8. Since syswrite()
refuses UTF8-flagged string containing characters whose code point is
greater than 255, it causes various errors in returning a response.